Change serveral table's node_id col's type
node_id in NodesTable, ClusterStateTable, ShardAuthoritativeStatus table
is BIGINT, which is different from the definition in AdminCommandTable.

This commit simply to make these table's node_id col's type to
INTEGER, for better consistency

+++ logdevice/ops/ldquery/tables/Nodes.cpp
@@ -26,7 +26,7 @@ namespace facebook {
       namespace tables {
 
 TableColumns Nodes::getColumns() const {
-  return {{"node_id", DataType::BIGINT, "Id of the node"},
+  return {{"node_id", DataType::INTEGER, "Id of the node"},
           {"name", DataType::TEXT, "Human readable name of the node"},
           {"address",
            DataType::TEXT,
